feat(runtime): enforce persistent run deadline Initialize absolute deadline in AgentRunContext metadata (default 3600s, MINDCLI_REACT_MAX_DURATION_SECONDS=0 disables) Persist deadline and inherit it across Plan tasks and Team children Block new LLM/tool calls and resume once deadline is expired or invalid Replace hard stagnation termination with advisory cycle warning using canonicalized tool request fingerprints Widen token counters from int to long Surface deadline status via /run inspect

+++ src/main/java/com/mindcli/capability/memory/MemoryManager.java
@@ -290,11 +290,11 @@ public class MemoryManager {
     /**
      * 记录 token 使用
      */
-    public void recordTokenUsage(int inputTokens, int outputTokens) {
+    public void recordTokenUsage(long inputTokens, long outputTokens) {
         tokenBudget.recordUsage(inputTokens, outputTokens);
     }
 
-    public void recordTokenUsage(int inputTokens, int outputTokens, int cachedInputTokens) {
+    public void recordTokenUsage(long inputTokens, long outputTokens, long cachedInputTokens) {
         tokenBudget.recordUsage(inputTokens, outputTokens, cachedInputTokens);
     }
