+++ src/main/java/com/mindcli/platform/llm/context/TokenUsageFormatter.java
@@ -9,11 +9,11 @@ public final class TokenUsageFormatter {
     private TokenUsageFormatter() {
     }
 
-    public static String format(LlmClient llmClient, int inputTokens, int outputTokens,
-                                int cachedInputTokens, long startNanos) {
+    public static String format(LlmClient llmClient, long inputTokens, long outputTokens,
+                                long cachedInputTokens, long startNanos) {
         ContextProfile profile = ContextProfile.from(llmClient);
         double elapsedSeconds = (System.nanoTime() - startNanos) / 1_000_000_000.0;
-        int total = Math.max(0, inputTokens) + Math.max(0, outputTokens);
+        long total = Math.max(0L, inputTokens) + Math.max(0L, outputTokens);
         String cost = estimatedCostCny(llmClient, inputTokens, outputTokens, cachedInputTokens);
         // 第 16 期：Y 显示 maxContextWindow（即模型窗口本身），不再用 80% × window 这种"软预算"
         // 误导用户。AgentBudget 默认已无硬限，撞窗口由 ConversationHistoryCompactor 自动压缩兜底。
@@ -21,14 +21,14 @@ public final class TokenUsageFormatter {
                 "📊 Token: 已用 %d / %d (cached: %d, 估算 %s) | 输入 %d / 输出 %d | ⏱ %.1fs",
                 total,
                 profile.maxContextWindow(),
-                Math.max(0, cachedInputTokens),
+                Math.max(0L, cachedInputTokens),
                 cost,
-                Math.max(0, inputTokens),
-                Math.max(0, outputTokens),
+                Math.max(0L, inputTokens),
+                Math.max(0L, outputTokens),
                 elapsedSeconds));
     }
 
-    public static String estimatedCostCny(LlmClient llmClient, int inputTokens, int outputTokens, int cachedInputTokens) {
+    public static String estimatedCostCny(LlmClient llmClient, long inputTokens, long outputTokens, long cachedInputTokens) {
         String provider = llmClient == null ? "" : llmClient.getProviderName();
         double inputPerMillion;
         double cachedPerMillion;
@@ -46,11 +46,11 @@ public final class TokenUsageFormatter {
             cachedPerMillion = 1.0;
             outputPerMillion = 15.0;
         }
-        int cached = Math.max(0, Math.min(inputTokens, cachedInputTokens));
-        int uncachedInput = Math.max(0, inputTokens - cached);
+        long cached = Math.max(0L, Math.min(inputTokens, cachedInputTokens));
+        long uncachedInput = Math.max(0L, inputTokens - cached);
         double cny = (uncachedInput / 1_000_000.0) * inputPerMillion
                 + (cached / 1_000_000.0) * cachedPerMillion
-                + (Math.max(0, outputTokens) / 1_000_000.0) * outputPerMillion;
+                + (Math.max(0L, outputTokens) / 1_000_000.0) * outputPerMillion;
         return String.format(Locale.ROOT, "¥%.4f", cny);
     }
 }
